zkDSL compiler: split the parser the same way
parser.rs was 1,581 lines and is now 861, with three modules beside it. Pure
moves: no digest changes, the guest is unchanged.

Split by the question each answers, and each doc states the rule its part lives
by. expr reads structure out of a line under one rule, that a string literal is
one opaque token and every scan goes through depth0; it also records that the
precedence there is correct as it stands, since that was measured rather than
assumed. consts evaluates a constant at parse time, which five syntactic
positions demand before lowering runs, and reads it as an INTEGER deliberately,
which is what makes a derived size right. subst substitutes an expression for a
name through a statement tree, which is how unroll and Const bind their
variable, and notes that a field dropped from one of its arms is a construct
that loses its meaning only inside a specialisation.

The two files that held nearly all of this crate are now nine modules, none over
900 lines.
