+++ crates/lean_compiler/src/parser.rs
@@ -1,8 +1,31 @@
 //! Parser: a minimal indentation-based Python-like surface syntax → [`Ast`].
+//!
+//! This file holds the line-oriented part: the indentation structure, the
+//! statement forms, and the top level's global constants. The rest is split by
+//! the question it answers:
+//!
+//! - [`mod@expr`] reads structure out of a line, under one rule: a string
+//!   literal is one opaque token, and every scan goes through `depth0`.
+//! - [`mod@consts`] evaluates a constant at parse time, which five syntactic
+//!   positions demand before lowering ever runs. It reads a constant as an
+//!   INTEGER, deliberately, which is what makes a derived size right.
+//! - [`mod@subst`] substitutes an expression for a name through a statement
+//!   tree, which is how `unroll` and `Const` bind their variable.
 
 use super::*;
 use std::collections::BTreeMap;
 
+mod consts;
+mod expr;
+mod subst;
+pub use consts::parse_const;
+use consts::{apply_replacements, const_int_expr, eval_const_int, gpow_bound, parse_f192_const, parse_gpow_bound};
+use expr::{
+    binding_name, call_args, is_ident, parse_expr, split_assign, split_aug, split_once_top, split_top, strip_comment,
+    strip_const_wrapper, string_lit, top_level_cmp,
+};
+pub(crate) use subst::subst_stmts;
+use subst::subst_var;
 /// Parse zkDSL source (the Python-shaped surface syntax of `zkDSL.md`) into an
 /// [`Ast`]. The `snark_lib` import is skipped; any other import is an error,
 /// since a program is a single file.
@@ -282,167 +305,6 @@ fn infer_return_shapes(funcs: &mut [Func]) -> Result<(), String> {
             .unwrap_or_else(|| vec![ReturnShape::Scalar; f.n_ret]);
     }
     Ok(())
-}
-
-fn parse_f192_const(s: &str) -> Option<Result<F192, String>> {
-    let inner = s.trim().strip_prefix("f192(")?.strip_suffix(')')?;
-    let parts = split_top(inner, ',');
-    Some((|| {
-        if parts.len() != 3 {
-            return Err("f192 needs exactly three limbs".into());
-        }
-        let mut limbs = [0u64; 3];
-        for (i, p) in parts.iter().enumerate() {
-            limbs[i] =
-                u64::try_from(eval_const_int(p.trim())?).map_err(|_| "an f192 limb does not fit in u64".to_string())?;
-        }
-        Ok(F192::new(limbs[0], limbs[1], limbs[2]))
-    })())
-}
-
-/// Apply identifier-level **placeholder** replacements to source text before
-/// parsing: each maximal run of identifier characters (`[A-Za-z0-9_]`) that
-/// equals a key of `replacements` is replaced by its value; other text,
-/// including substrings of longer identifiers, is untouched. Mirrors leanVM's
-/// `CompilationFlags::replacements`. An empty map returns the source unchanged.
-fn apply_replacements(src: &str, replacements: &BTreeMap<String, String>) -> String {
-    if replacements.is_empty() {
-        return src.to_string();
-    }
-    let is_ident_char = |c: char| c.is_alphanumeric() || c == '_';
-    let mut out = String::with_capacity(src.len());
-    let mut word = String::new(); // current run of identifier characters
-    let flush = |out: &mut String, word: &mut String| {
-        match replacements.get(word.as_str()) {
-            Some(v) => out.push_str(v),
-            None => out.push_str(word),
-        }
-        word.clear();
-    };
-    for c in src.chars() {
-        if is_ident_char(c) {
-            word.push(c);
-        } else {
-            flush(&mut out, &mut word);
-            out.push(c);
-        }
-    }
-    flush(&mut out, &mut word);
-    out
-}
-
-/// `raw` without its trailing comment. A `#` inside a string literal is part of
-/// the string: truncating there dropped the rest of the line, and the shortened
-/// line usually still parsed.
-fn strip_comment(raw: &str) -> &str {
-    let mut in_str = false;
-    for (i, c) in raw.char_indices() {
-        match c {
-            '"' => in_str = !in_str,
-            '#' if !in_str => return &raw[..i],
-            _ => {}
-        }
-    }
-    raw
-}
-
-/// The inside of a `const(...)` wrapping the WHOLE of `s`, or `None`.
-///
-/// `const(a) == b` is not one: its first `)` closes before the end, so the
-/// wrapper is a subterm and the condition as a whole is an ordinary one.
-fn strip_const_wrapper(s: &str) -> Option<&str> {
-    let inner = s.trim().strip_prefix("const(")?.strip_suffix(')')?;
-    let mut depth = 0i32;
-    for c in inner.bytes() {
-        match c {
-            b'(' | b'[' => depth += 1,
-            b')' | b']' => {
-                depth -= 1;
-                if depth < 0 {
-                    return None;
-                }
-            }
-            _ => {}
-        }
-    }
-    Some(inner)
-}
-
-/// The first top-level comparison operator in `s`. Used only on a line that
-/// reached the bare-call fallback, so an `assert` or an assignment never gets
-/// here and a comparison nested in a call sits at depth > 0.
-fn top_level_cmp(s: &str) -> Option<&'static str> {
-    let b = s.as_bytes();
-    for (i, c) in depth0(s) {
-        let eq = b.get(i + 1) == Some(&b'=');
-        match c {
-            b'=' if eq => return Some("=="),
-            b'!' if eq => return Some("!="),
-            b'<' => return Some(if eq { "<=" } else { "<" }),
-            b'>' => return Some(if eq { ">=" } else { ">" }),
-            _ => {}
-        }
-    }
-    None
-}
-
-/// A binding or parameter name, validated. Anything else reaching here is
-/// either a mis-split (`x /`, `x !`) or a top-level constant substituted into a
-/// binding position: constants are replaced textually before parsing, so
-/// `V = 8` followed by `def scale(V)` arrives as a parameter literally named
-/// `8`, and the body's `V` reads the constant. That compiled, and returned 16
-/// for `scale(3)`. `zkDSL.md` §Global constants reserves the name; this is what
-/// enforces it.
-fn binding_name(raw: &str, what: &str) -> Result<String, String> {
-    let n = raw.trim();
-    if is_ident(n) {
-        return Ok(n.to_string());
-    }
-    Err(format!(
-        "`{n}` is not a valid {what}: a name must be a plain identifier. A top-level constant's name is \
-         reserved, and is substituted before parsing, so a parameter or local may not reuse one."
-    ))
-}
-
-/// A plain identifier: non-empty, starts with a letter or `_`, all
-/// `[A-Za-z0-9_]` (no operators, brackets, or commas).
-fn is_ident(s: &str) -> bool {
-    let mut cs = s.chars();
-    matches!(cs.next(), Some(c) if c.is_alphabetic() || c == '_') && s.chars().all(|c| c.is_alphanumeric() || c == '_')
-}
-
-/// Evaluate a compile-time **integer** constant expression: decimal literals
-/// combined with `+ - * / // % **` and parentheses. This is ordinary integer
-/// arithmetic (a global constant is a count, a size, an exponent), deliberately
-/// distinct from runtime field arithmetic, so a derived size like `2 + (W - 1) *
-/// V + LOG_LIFETIME` comes out right; a single `/` divides like `//` here.
-/// References to earlier constants are already substituted to their decimal
-/// values, so the input is pure arithmetic. Overflow, division by zero, and a
-/// negative intermediate are errors.
-fn eval_const_int(s: &str) -> Result<u128, String> {
-    parse_expr(s)
-        .ok()
-        .and_then(|e| const_int_expr(&e))
-        .ok_or_else(|| format!("not a compile-time integer constant expression: `{}`", s.trim()))
-}
-
-/// Evaluate a compile-time constant expression (integer literals, `GEN`,
-/// `GEN ** k`, and `+`/`*` combinations of those) to its field element.
-/// Used for the `# public_input: <elt>, <elt>` annotation of `.py` test
-/// programs (see `tests/py_source.rs`).
-pub fn parse_const(s: &str) -> Result<F192, String> {
-    fn eval(e: &Expr) -> Result<F192, String> {
-        match e {
-            // An integer literal is the raw 128-bit bit pattern of a machine word.
-            Expr::Lit(n) => Ok(F192::new(*n as u64, (*n >> 64) as u64, 0)),
-            Expr::Gen => Ok(g_pow(1).into()),
-            Expr::GPow(k) => Ok(g_pow_u128(*k).into()),
-            Expr::Add(a, b) => Ok(eval(a)? + eval(b)?),
-            Expr::Mul(a, b) => Ok(eval(a)? * eval(b)?),
-            other => Err(format!("not a constant expression: `{other:?}`")),
-        }
-    }
-    eval(&parse_expr(s)?)
 }
 
 /// Parse a zkDSL source file (a `.py` file, since the DSL is Python-shaped, see
@@ -952,414 +814,13 @@ impl Parser {
     }
 }
 
-/// The bytes of `s` that sit outside every `(…)` / `[…]` group, with their
-/// index: the one scanner behind all the top-level splits below. Brackets
-/// themselves are never yielded, so a separator that is a bracket never splits.
-fn depth0(s: &str) -> impl Iterator<Item = (usize, u8)> + '_ {
-    let mut depth = 0i32;
-    let mut in_str = false;
-    s.as_bytes().iter().enumerate().filter_map(move |(i, &c)| {
-        // A string literal is one opaque token. Without this every splitter
-        // below reads the brackets and operators SPELLED INSIDE a stream name as
-        // structure: `hint_witness(b, "a,b")` split into three arguments.
-        if in_str {
-            in_str = c != b'"';
-            return None;
-        }
-        match c {
-            b'"' => {
-                in_str = true;
-                None
-            }
-            b'(' | b'[' => {
-                depth += 1;
-                None
-            }
-            b')' | b']' => {
-                depth -= 1;
-                None
-            }
-            _ => (depth == 0).then_some((i, c)),
-        }
-    })
-}
-
 type Aug = Option<(String, &'static str, String)>;
-
-/// A top-level augmented assignment `lhs OP= rhs` -> `(lhs, "OP", rhs)`, for
-/// OP in `+ - * // %`. `Ok(None)` for a plain `=` or a comparison (`==`, `!=`,
-/// `<=`, `>=`); an `Err` for a compound spelling this language does not have.
-/// The operator's `=` must sit at depth 0 and be immediately preceded by
-/// exactly the operator characters.
-///
-/// The unsupported spellings must be REJECTED rather than declined: falling
-/// through left [`split_assign`] to split the bare `=`, so `x /= 2` became a
-/// binding named `x /` and the program silently kept the old `x`.
-fn split_aug(s: &str) -> Result<Aug, String> {
-    let b = s.as_bytes();
-    for (i, c) in depth0(s) {
-        if c != b'=' {
-            continue;
-        }
-        // Nothing to the left is no assignment at all; `split_assign` and the
-        // binding-name check below give that its error.
-        if i == 0 {
-            return Ok(None);
-        }
-        // not `==` and not a comparison tail (`<=`, `>=`, `!=`)
-        if b.get(i + 1) == Some(&b'=') || matches!(b[i - 1], b'=' | b'<' | b'>' | b'!') {
-            return Ok(None);
-        }
-        let prev2 = b.get(i.wrapping_sub(2)).copied();
-        let (op, plen): (&str, usize) = match b[i - 1] {
-            b'+' => ("+", 1),
-            b'-' => ("-", 1),
-            b'%' => ("%", 1),
-            b'*' if prev2 == Some(b'*') => {
-                return Err("`**=` is not supported; write `x = x ** k`".into());
-            }
-            b'*' => ("*", 1),
-            b'/' if prev2 == Some(b'/') => ("//", 2),
-            b'/' => {
-                return Err(
-                    "`/=` is not supported; `/` is runtime field division, so write `x = x / y` (or `//=` for the \
-                     compile-time floor division)"
-                        .into(),
-                );
-            }
-            _ => return Ok(None),
-        };
-        return Ok(Some((
-            s[..i - plen].trim().to_string(),
-            op,
-            s[i + 1..].trim().to_string(),
-        )));
-    }
-    Ok(None)
-}
-
-/// Split on a top-level BARE `=`: not part of `==`, not the tail of a
-/// comparison (`!=`, `<=`, `>=`), and not the tail of a compound assignment
-/// ([`split_aug`] owns those, and rejects the ones this language lacks).
-/// Without the last two exclusions a bare `x != y` split into a binding named
-/// `x !`, which in a verifier is an `assert` that compiled to nothing.
-fn split_assign(s: &str) -> Option<(String, String)> {
-    let b = s.as_bytes();
-    for (i, c) in depth0(s) {
-        if c != b'=' || b.get(i + 1) == Some(&b'=') {
-            continue;
-        }
-        if i > 0 && matches!(b[i - 1], b'=' | b'<' | b'>' | b'!' | b'+' | b'-' | b'*' | b'/' | b'%') {
-            continue;
-        }
-        return Some((s[..i].to_string(), s[i + 1..].to_string()));
-    }
-    None
-}
-
-/// Split `s` on every top-level occurrence of the ASCII char `sep`.
-fn split_top(s: &str, sep: char) -> Vec<String> {
-    let sep = sep as u8;
-    let mut parts = Vec::new();
-    let mut start = 0;
-    for (i, c) in depth0(s) {
-        if c == sep {
-            parts.push(s[start..i].to_string());
-            start = i + 1;
-        }
-    }
-    parts.push(s[start..].to_string());
-    parts
-}
-
-/// Split `s` at the top-level additive tier: operands and the `+` / `-`
-/// operators between them. Left-associative; parenthesised/bracketed sub-terms
-/// are left intact.
-fn split_add(s: &str) -> (Vec<String>, Vec<u8>) {
-    let (mut segs, mut ops) = (Vec::new(), Vec::new());
-    let mut start = 0usize;
-    for (i, c) in depth0(s) {
-        if c == b'+' || c == b'-' {
-            segs.push(s[start..i].to_string());
-            ops.push(c);
-            start = i + 1;
-        }
-    }
-    segs.push(s[start..].to_string());
-    (segs, ops)
-}
-
-/// Split `s` at the top-level multiplicative tier: the operands and the
-/// operators between them (`*`, `//` for floor-division, `/` for runtime field
-/// division, `%` for remainder). A `**` power is left intact (bound tighter).
-/// Left-associative.
-fn split_mul(s: &str) -> (Vec<String>, Vec<u8>) {
-    let b = s.as_bytes();
-    let (mut segs, mut ops) = (Vec::new(), Vec::new());
-    let (mut start, mut next) = (0usize, 0usize);
-    for (i, c) in depth0(s) {
-        if i < next {
-            continue; // the second `/` of a `//`, already consumed
-        }
-        let (op, len) = match c {
-            b'*' if b.get(i + 1) == Some(&b'*') || (i > 0 && b[i - 1] == b'*') => continue, // `**`
-            b'*' => (b'*', 1),
-            b'/' if b.get(i + 1) == Some(&b'/') => (b'/', 2), // `//` compile-time floor-division
-            b'/' => (b'd', 1),                                // `/` runtime field division
-            b'%' => (b'%', 1),
-            _ => continue,
-        };
-        segs.push(s[start..i].to_string());
-        ops.push(op);
-        next = i + len;
-        start = next;
-    }
-    segs.push(s[start..].to_string());
-    (segs, ops)
-}
-
-/// Split `s` once on a top-level multi-char operator `op`.
-fn split_once_top(s: &str, op: &str) -> Option<(String, String)> {
-    let b = s.as_bytes();
-    for (i, _) in depth0(s) {
-        if b[i..].starts_with(op.as_bytes()) {
-            return Some((s[..i].to_string(), s[i + op.len()..].to_string()));
-        }
-    }
-    None
-}
-
-/// The top-level arguments of a `name(a, b, …)` call, or `None` when `line` is
-/// not one. Zero arguments come back as one empty string, as [`split_top`]
-/// gives them.
-fn call_args(line: &str, name: &str) -> Option<Vec<String>> {
-    let inner = line.trim().strip_prefix(name)?.strip_prefix('(')?.strip_suffix(')')?;
-    Some(split_top(inner, ','))
-}
-
-/// The contents of a `"…"` string literal.
-fn string_lit(s: &str) -> Option<&str> {
-    s.trim().strip_prefix('"')?.strip_suffix('"')
-}
-
-/// Fold a compile-time INTEGER expression (literals combined with the usual
-/// operators) to its value; `None` if any leaf is not a literal. Placeholders
-/// are substituted before parsing, so `GEN ** (K_SKIP + 1)`-style exponents
-/// fold here.
-fn const_int_expr(e: &Expr) -> Option<u128> {
-    match e {
-        Expr::Lit(k) => Some(*k),
-        Expr::Add(a, b) => const_int_expr(a)?.checked_add(const_int_expr(b)?),
-        Expr::Sub(a, b) => const_int_expr(a)?.checked_sub(const_int_expr(b)?),
-        Expr::Mul(a, b) => const_int_expr(a)?.checked_mul(const_int_expr(b)?),
-        // A single `/` between compile-time integers is integer division: in a
-        // constant (a count, a size), there is no field to divide in.
-        Expr::Div(a, b) | Expr::FieldDiv(a, b) => match const_int_expr(b)? {
-            0 => None,
-            d => Some(const_int_expr(a)? / d),
-        },
-        Expr::Mod(a, b) => match const_int_expr(b)? {
-            0 => None,
-            d => Some(const_int_expr(a)? % d),
-        },
-        Expr::Pow(a, b) => const_int_expr(a)?.checked_pow(u32::try_from(const_int_expr(b)?).ok()?),
-        _ => None,
-    }
-}
-
-/// A range bound (`mul_range` bounds and `assert log _ < log _` bounds): a
-/// compile-time power of the generator (`1` = `g^0`, `GEN` = `g^1`, or
-/// `GEN ** k`), returning the exponent `k`. Both uses walk/compare exponents,
-/// so the bound must name `g^k` explicitly (an element that is not a known
-/// power of `g` has no usable exponent).
-fn gpow_bound(e: &Expr) -> Result<u64, String> {
-    match e {
-        // `g` is `x`, so the literal `2^k` IS `g^k`, and `1` is `g^0`. Rejecting
-        // these used to make `mul_range(1, 8)` an error although it runs exactly
-        // like `mul_range(1, GEN ** 3)`, and `mul_range(2, GEN ** 5)` an error
-        // although `2 == GEN`. It also contradicted `gaddr_of`, which reads the
-        // same literal as the same element.
-        Expr::Lit(n) if (n.is_power_of_two() && *n < (1 << 64)) || *n == 1 => Ok(n.trailing_zeros() as u64),
-        Expr::Gen => Ok(1),
-        Expr::GPow(k) => u64::try_from(*k).map_err(|_| format!("bound exponent {k} does not fit in u64")),
-        other => Err(format!(
-            "a range bound must be a power of GEN (`1`, `GEN`, or `GEN ** k`), got `{other:?}`"
-        )),
-    }
-}
-
-fn parse_gpow_bound(s: &str) -> Result<u64, String> {
-    gpow_bound(&parse_expr(s)?)
-}
 
 /// Strip a leading `log` token (`log x`, `log(x)`), if present. The token must
 /// end at a boundary, so a variable named `logx` is not a log of `x`.
 fn strip_log(s: &str) -> Option<&str> {
     let r = s.trim_start().strip_prefix("log")?;
     r.starts_with([' ', '(']).then_some(r)
-}
-
-/// Substitute `Var(name)` → `to` through a statement list, stopping at a
-/// statement that rebinds `name` (later uses refer to the new binding).
-/// Nested blocks recurse independently, since their bindings are branch-local,
-/// matching the lowering's scoping. Used by `Const`-parameter specialization.
-pub(crate) fn subst_stmts(stmts: &[Stmt], name: &str, to: &Expr) -> Vec<Stmt> {
-    let mut out = Vec::with_capacity(stmts.len());
-    let mut active = true;
-    for s in stmts {
-        if !active {
-            out.push(s.clone());
-            continue;
-        }
-        let (s, rebinds) = subst_stmt(s, name, to);
-        out.push(s);
-        active = !rebinds;
-    }
-    out
-}
-
-/// One statement of [`subst_stmts`]; the flag says whether it rebinds `name`.
-fn subst_stmt(s: &Stmt, name: &str, to: &Expr) -> (Stmt, bool) {
-    let (kind, rebinds) = subst_kind(&s.kind, name, to);
-    (s.at(kind), rebinds)
-}
-
-/// The kind half of [`subst_stmt`]; the flag says whether it rebinds `name`.
-fn subst_kind(s: &StmtKind, name: &str, to: &Expr) -> (StmtKind, bool) {
-    let e = |x: &Expr| subst_var(x, name, to);
-    match s {
-        StmtKind::Let(n, x) => (StmtKind::Let(n.clone(), e(x)), n == name),
-        StmtKind::LetTuple(ns, f, args) => (
-            StmtKind::LetTuple(ns.clone(), f.clone(), args.iter().map(e).collect()),
-            ns.iter().any(|n| n == name),
-        ),
-        StmtKind::AssertEq(a, b) => (StmtKind::AssertEq(e(a), e(b)), false),
-        StmtKind::AssertNe(a, b) => (StmtKind::AssertNe(e(a), e(b)), false),
-        StmtKind::AssertLt(a, bound) => (
-            StmtKind::AssertLt(
-                e(a),
-                match bound {
-                    LtBound::Const(k) => LtBound::Const(*k),
-                    LtBound::Runtime(b) => LtBound::Runtime(e(b)),
-                },
-            ),
-            false,
-        ),
-        StmtKind::Call(f, args) => (StmtKind::Call(f.clone(), args.iter().map(e).collect()), false),
-        StmtKind::Print { label, value } => (
-            StmtKind::Print {
-                label: label.clone(),
-                value: e(value),
-            },
-            false,
-        ),
-        StmtKind::HintWitness { dest, name: n } => (
-            StmtKind::HintWitness {
-                dest: e(dest),
-                name: n.clone(),
-            },
-            false,
-        ),
-        StmtKind::Store(a, i, v) => (StmtKind::Store(e(a), e(i), e(v)), false),
-        StmtKind::Return(es) => (StmtKind::Return(es.iter().map(e).collect()), false),
-        StmtKind::CallIfNe(a, b, f, args) => (
-            StmtKind::CallIfNe(e(a), e(b), f.clone(), args.iter().map(e).collect()),
-            false,
-        ),
-        StmtKind::For { var, lo, hi, body } => {
-            let hi = match hi {
-                ForBound::Const(k) => ForBound::Const(*k),
-                ForBound::Runtime(b) => ForBound::Runtime(e(b)),
-            };
-            // The counter shadows `name` inside the body only.
-            let body = if var == name {
-                body.clone()
-            } else {
-                subst_stmts(body, name, to)
-            };
-            (
-                StmtKind::For {
-                    var: var.clone(),
-                    lo: *lo,
-                    hi,
-                    body,
-                },
-                false,
-            )
-        }
-        StmtKind::Unroll { var, lo, hi, body } => {
-            let body = if var == name {
-                body.clone()
-            } else {
-                subst_stmts(body, name, to)
-            };
-            (
-                StmtKind::Unroll {
-                    var: var.clone(),
-                    lo: e(lo),
-                    hi: e(hi),
-                    body,
-                },
-                false,
-            )
-        }
-        StmtKind::If {
-            eq,
-            lhs,
-            rhs,
-            then,
-            els,
-            force_const,
-        } => (
-            StmtKind::If {
-                eq: *eq,
-                lhs: e(lhs),
-                rhs: e(rhs),
-                then: subst_stmts(then, name, to),
-                els: subst_stmts(els, name, to),
-                force_const: *force_const,
-            },
-            false,
-        ),
-        StmtKind::Match { x, cases } => (
-            StmtKind::Match {
-                x: e(x),
-                cases: cases.iter().map(|c| subst_stmts(c, name, to)).collect(),
-            },
-            false,
-        ),
-        StmtKind::LetMatchRange { names, x, arms } => (
-            StmtKind::LetMatchRange {
-                names: names.clone(),
-                x: e(x),
-                arms: arms.iter().map(e).collect(),
-            },
-            names.iter().any(|n| n == name),
-        ),
-    }
-}
-
-/// `e` with every `Var(name)` replaced by `to`: the `match_range` arm
-/// expansion, where the lambda parameter becomes the arm's integer literal.
-fn subst_var(e: &Expr, name: &str, to: &Expr) -> Expr {
-    let s = |b: &Expr| Box::new(subst_var(b, name, to));
-    match e {
-        Expr::Var(v) if v == name => to.clone(),
-        Expr::Add(a, b) => Expr::Add(s(a), s(b)),
-        Expr::Mul(a, b) => Expr::Mul(s(a), s(b)),
-        Expr::Sub(a, b) => Expr::Sub(s(a), s(b)),
-        Expr::Div(a, b) => Expr::Div(s(a), s(b)),
-        Expr::FieldDiv(a, b) => Expr::FieldDiv(s(a), s(b)),
-        Expr::Mod(a, b) => Expr::Mod(s(a), s(b)),
-        Expr::Index(a, b) => Expr::Index(s(a), s(b)),
-        Expr::Slice(a, lo, hi) => Expr::Slice(s(a), s(lo), s(hi)),
-        Expr::GenPow(e) => Expr::GenPow(s(e)),
-        Expr::Pow(a, b) => Expr::Pow(s(a), s(b)),
-        Expr::HeapBufDyn(sz) => Expr::HeapBufDyn(s(sz)),
-        Expr::ListLit(es) => Expr::ListLit(es.iter().map(|a| subst_var(a, name, to)).collect()),
-        Expr::Call(f, args) => Expr::Call(f.clone(), args.iter().map(|a| subst_var(a, name, to)).collect()),
-        other => other.clone(),
-    }
 }
 
 /// `names = match_range(log(x), range(a, b), lambda i: expr, …)`: leanVM's
@@ -1409,173 +870,4 @@ fn parse_match_range(lhs: &str, rhs: &str) -> Result<StmtKind, String> {
         }
     }
     Ok(StmtKind::LetMatchRange { names, x, arms })
-}
-
-/// Combine one tier's operands left-associatively: `node` builds the AST node
-/// for each operator (as [`split_add`] / [`split_mul`] tag it).
-fn fold_ops(segs: &[String], ops: &[u8], node: impl Fn(u8, Box<Expr>, Box<Expr>) -> Expr) -> Result<Expr, String> {
-    // An empty operand is an operator missing a side: a leading `-`, a trailing
-    // operator, or two in a row. Naming it beats letting `parse_expr("")` report
-    // an empty backtick, which is what every one of these used to say.
-    if let Some(i) = segs.iter().position(|seg| seg.trim().is_empty()) {
-        let (op, side) = if i == 0 { (ops[0], "left") } else { (ops[i - 1], "right") };
-        // `split_mul` encodes the two divisions, so spell them back out.
-        let shown = match op {
-            b'/' => "//".to_string(),
-            b'd' => "/".to_string(),
-            c => (c as char).to_string(),
-        };
-        let hint = if op == b'-' && i == 0 {
-            ": there is no unary minus, and field subtraction is `+`"
-        } else {
-            ""
-        };
-        return Err(format!("`{shown}` has no {side} operand{hint}"));
-    }
-    let mut acc = parse_expr(&segs[0])?;
-    for (&op, seg) in ops.iter().zip(&segs[1..]) {
-        let rhs = Box::new(parse_expr(seg)?);
-        acc = node(op, Box::new(acc), rhs);
-    }
-    Ok(acc)
-}
-
-/// Parse an expression with `+` (lowest) then `*`, atoms being integer literals,
-/// variables, calls `f(args)`, and parenthesised sub-expressions.
-fn parse_expr(s: &str) -> Result<Expr, String> {
-    let s = s.trim();
-    // `+` / `-` at top level (lowest precedence), left-associative. `-` is
-    // compile-time integer subtraction (field subtraction is `+` = XOR).
-    let (segs, ops) = split_add(s);
-    if !ops.is_empty() {
-        return fold_ops(&segs, &ops, |op, l, r| match op {
-            b'+' => Expr::Add(l, r),
-            _ => Expr::Sub(l, r),
-        });
-    }
-    // `*`, `/`, `//`, `%` (bind tighter than `+`), skipping the two-char `**`.
-    let (segs, ops) = split_mul(s);
-    if !ops.is_empty() {
-        return fold_ops(&segs, &ops, |op, l, r| match op {
-            b'*' => Expr::Mul(l, r),
-            b'/' => Expr::Div(l, r),
-            b'd' => Expr::FieldDiv(l, r),
-            _ => Expr::Mod(l, r),
-        });
-    }
-    // `**` (compile-time power), tightest binding: `base ** k` with `k` an
-    // integer literal (possibly large), or a parenthesised compile-time
-    // integer expression like `GEN ** (2 * s + 1)`, evaluated at lowering,
-    // so it can reference `unroll` counters and constants.
-    if let Some((base, exp)) = split_once_top(s, "**") {
-        let base = parse_expr(&base)?;
-        let exp_e = parse_expr(&exp)?;
-        return match base {
-            // `GEN ** k`: a compile-time integer exponent (a literal or a
-            // constant expression like `K_SKIP + 1`) folds to `g^k`; a runtime
-            // expression (e.g. an `unroll` var) becomes `GenPow`, resolved at
-            // lowering.
-            Expr::Gen => match const_int_expr(&exp_e) {
-                Some(k) => Ok(Expr::GPow(k)),
-                None => Ok(Expr::GenPow(Box::new(exp_e))),
-            },
-            // Any other base with a compile-time exponent: square-and-multiply.
-            _ => Ok(Expr::Pow(Box::new(base), Box::new(exp_e))),
-        };
-    }
-    // Atom.
-    if s.starts_with('(') && s.ends_with(')') {
-        return parse_expr(&s[1..s.len() - 1]);
-    }
-    if s == "GEN" {
-        return Ok(Expr::Gen);
-    }
-    if let Ok(n) = s.parse::<u128>() {
-        return Ok(Expr::Lit(n));
-    }
-    // List literal `[a, b, …]`: an initialized StackBuf (only meaningful as
-    // the RHS of an assignment; top-level constant arrays are parsed earlier).
-    if s.starts_with('[') && s.ends_with(']') {
-        let inner = s[1..s.len() - 1].trim();
-        if inner.is_empty() {
-            return Err("a list literal needs at least one element".into());
-        }
-        return Ok(Expr::ListLit(
-            split_top(inner, ',')
-                .iter()
-                .map(|e| parse_expr(e))
-                .collect::<Result<_, _>>()?,
-        ));
-    }
-    // Index `base[idx]` or slice `base[lo:hi]` (binds tightest, like a call).
-    if s.ends_with(']') {
-        let open = s.find('[').ok_or_else(|| format!("unbalanced `]` in `{s}`"))?;
-        let base = parse_expr(&s[..open])?;
-        let inner = &s[open + 1..s.len() - 1];
-        if let Some((lo, hi)) = split_once_top(inner, ":") {
-            return Ok(Expr::Slice(
-                Box::new(base),
-                Box::new(parse_expr(&lo)?),
-                Box::new(parse_expr(&hi)?),
-            ));
-        }
-        let idx = parse_expr(inner)?;
-        return Ok(Expr::Index(Box::new(base), Box::new(idx)));
-    }
-    if let Some(open) = s.find('(')
-        && s.ends_with(')')
-    {
-        let name = s[..open].trim().to_string();
-        let args_str = s[open + 1..s.len() - 1].trim();
-        let args = if args_str.is_empty() {
-            vec![]
-        } else {
-            split_top(args_str, ',')
-                .iter()
-                .map(|a| {
-                    if let Some((key, value)) = split_once_top(a, "=") {
-                        let key = key.trim();
-                        if key.is_empty() || !key.chars().all(|c| c.is_ascii_alphanumeric() || c == '_') {
-                            return Err(format!("invalid keyword argument `{key}`"));
-                        }
-                        Ok(Expr::Call(format!("__kw_{key}"), vec![parse_expr(&value)?]))
-                    } else {
-                        parse_expr(a)
-                    }
-                })
-                .collect::<Result<_, _>>()?
-        };
-        // `HeapBuf(n)` / `StackBuf(n)` are allocations, not ordinary calls. A size
-        // that folds as parse-time integer arithmetic (`MAXQ + 1`, constants
-        // already substituted) is a static size like a bare literal.
-        // A cell count is a frame/heap size: reject one that does not fit rather
-        // than wrapping it into a plausible small buffer.
-        let cells = |n: u128| u64::try_from(n).map_err(|_| format!("{name} size {n} does not fit in u64"));
-        if name == "HeapBuf" {
-            if let Ok(n) = eval_const_int(args_str) {
-                return Ok(Expr::HeapBuf(cells(n)?));
-            }
-            return match args.as_slice() {
-                // A literal size is baked into the bytecode; any other
-                // expression is a runtime size (its low word is the count).
-                [Expr::Lit(n)] => Ok(Expr::HeapBuf(cells(*n)?)),
-                [e] => Ok(Expr::HeapBufDyn(Box::new(e.clone()))),
-                _ => Err("HeapBuf(size) takes one argument".into()),
-            };
-        }
-        if name == "StackBuf" {
-            if let Ok(n) = eval_const_int(args_str) {
-                return Ok(Expr::StackBuf(cells(n)?));
-            }
-            return Err("StackBuf(n) needs a parse-time integer size".into());
-        }
-        return Ok(Expr::Call(name, args));
-    }
-    if s.chars().all(|c| c.is_alphanumeric() || c == '_') && !s.is_empty() {
-        return Ok(Expr::Var(s.to_string()));
-    }
-    if s.is_empty() {
-        return Err("expected an expression".into());
-    }
-    Err(format!("cannot parse expression `{s}`"))
 }
